Clear static analysis warnings
Fixing static analysis warnings: adding assert for nonnull pointer;
Casting to 'unsigned int' before shift; initializing local variable

BUG=webm:2312

+++ av1/encoder/hash_motion.c
@@ -393,7 +393,8 @@ void av1_get_block_hash_value(uint8_t *y_src, int stride, int block_size,
                               uint32_t *hash_value1, uint32_t *hash_value2,
                               int use_highbitdepth, MACROBLOCK *x) {
   uint32_t to_hash[4];
-  const int add_value = hash_block_size_to_index(block_size) << crc_bits;
+  const int add_value =
+      (int)((unsigned int)hash_block_size_to_index(block_size) << crc_bits);
   assert(add_value >= 0);
   const int crc_mask = (1 << crc_bits) - 1;
